requirements.txt
azure-core==1.9.0
azure-eventhub==5.2.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
six==1.15.0
uamqp==1.2.12
urllib3==1.26.2
